+++ src/main/java/com/rahu/springjwt/payload/response/ProductHistoryResponse.java
@@ -18,7 +18,7 @@ public class ProductHistoryResponse {
   private Category category;
   private String name;
   private String description;
-  private Float price;
+  private Double price;
   private boolean images;
   private List<ProductHistoryDto> productHistoryContent;
   private List<ProductHistoryDto> prodHisContent;
@@ -27,7 +27,7 @@ public class ProductHistoryResponse {
   private long totalitems;
   private int totalpages;
 
-  public ProductHistoryResponse(String token, Long id, Category category, String name, String description, Float price,
+  public ProductHistoryResponse(String token, Long id, Category category, String name, String description, Double price,
                                 boolean images) {
     this.token = token;
     this.id = id;
