+++ src/main/java/com/rahu/springjwt/payload/response/ReportResponse.java
@@ -23,7 +23,7 @@ public class ReportResponse {
   private Category category;
   private String name;
   private String description;
-  private Float price;
+  private Double price;
   private boolean images;
   private List<ProductDto> list;
   private List<ProductOrderInvoiceDto> productOrderInvoiceDtos;
@@ -32,7 +32,7 @@ public class ReportResponse {
   private long totalitems;
   private int totalpages;
 
-  public ReportResponse(String token, Long id, Category category, String name, String description, Float price,
+  public ReportResponse(String token, Long id, Category category, String name, String description, Double price,
                         boolean images) {
     this.token = token;
     this.id = id;
